fix: resolve circular dependency by extracting router to separate file
